requirements.txt
# Core Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# LLM & Agent Orchestration
langgraph>=0.0.1
langchain>=0.1.0
openai>=1.0.0

# API Integrations (Placeholder - Update with real keys)
# tavily-python>=0.1.0  # Tavily AI
# nylas>=6.0.0  # Nylas Calendar
# google-maps-services>=1.6.3  # Google Maps

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Code Quality
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0
nylas==6.14.2
faster-whisper==1.2.1
# Optional for production
# redis>=5.0.0  # Session state management
# gunicorn>=21.0.0  # Production server
